requirements.txt
pydantic
colorlog
tenacity
dumb-menu
ujson
ruamel.yaml
tqdm
pythonnet
aiohttp
UnityPy